requirements.txt
streamlit==1.50.0
pandas==2.3.3
numpy==2.3.4
plotly==6.3.1
langgraph==1.0.0
langchain==1.0.1
langchain-groq==1.0.0
groq==0.32.0
scikit-learn==1.7.2
xgboost==3.1.0
shap==0.49.1
python-dotenv==1.1.1
streamlit-chat==0.1.1
streamlit-option-menu==0.4.0
streamlit-aggrid==1.1.9
langfuse>=2.0.0
